+++ 1_foundations/community_contributions/amirna2_contributions/personal-ai/models/__init__.py
@@ -1,0 +1,18 @@
+"""Model exports for the career chatbot.
+
+This package separates data models from the main chatbot implementation
+to keep `career_chatbot.py` focused on orchestration and logic.
+"""
+
+from .config import ChatbotConfig
+from .evaluation import Evaluation
+from .responses import StructuredResponse
+from .job_match import SkillAssessment, JobMatchResult
+
+__all__ = [
+    "ChatbotConfig",
+    "Evaluation",
+    "StructuredResponse",
+    "SkillAssessment",
+    "JobMatchResult",
+]
