+++ 1_foundations/community_contributions/amirna2_contributions/personal-ai/models/evaluation.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Evaluation(BaseModel):
+    """Evaluation result for a response."""
+    is_acceptable: bool
+    feedback: str
