+++ 1_foundations/community_contributions/amirna2_contributions/personal-ai/models/responses.py
@@ -1,0 +1,10 @@
+from typing import List
+from pydantic import BaseModel
+
+
+class StructuredResponse(BaseModel):
+    """Structured response with reasoning and evidence."""
+    response: str
+    reasoning: str
+    tools_used: List[str]
+    facts_used: List[str]
